feat(paths): add shared path normalization for review layer
